update deps & cleanup test setup
