+++ src/config.rs
@@ -212,15 +212,7 @@ mod tests {
         let config = MiasmaConfig {
             port: 8080,
             host: "127.0.0.1".to_string(),
-            #[cfg(unix)]
-            unix_socket: None,
-            max_in_flight: 100,
-            link_prefix: LinkPrefix::from_str("/").unwrap(),
-            link_count: 5,
-            max_depth: MaxDepth(None),
-            force_gzip: false,
-            unsafe_allow_html: false,
-            poison_source: Url::parse("https://example.com").unwrap(),
+            ..MiasmaConfig::default()
         };
 
         assert_eq!(config.address(), "127.0.0.1:8080");
